style: main화면 상단 부분 css
