+++ src/components/Main/SelectFriend.css
@@ -1,0 +1,28 @@
+.select-friend-container {
+  width: 100%;
+  height: 70px;
+  padding: 40px 16px;
+  background: rgb(250, 250, 250);
+}
+
+.select-friend-content {
+  width: 100%;
+  max-width: 1024px;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.select-friend-image {
+  width: 70px;
+  height: 70px;
+  border-radius: 24px;
+}
+
+.select-friend-text {
+  padding-left: 16px;
+  font-size: 28px;
+  line-height: 35px;
+  color: #333;
+}
