+++ src/components/Main/SelectTheme/SelectTheme.tsx
@@ -1,9 +1,80 @@
-import { ThemeGrid } from './ThemeGrid';
+import './SelectTheme.css';
+
+import { Link } from 'react-router-dom';
+
+import { ThemeItem } from './ThemeItem';
+
+const themes = [
+  {
+    title: '생일',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '졸업선물',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '스몰럭셔리',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '명품선물',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '결혼/집들이',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '따뜻한선물',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '가벼운선물',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '팬심저격',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '교환권',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '건강/비타민',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '과일/한우',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+  {
+    title: '출산/키즈',
+    imgUrl:
+      'https://img1.daumcdn.net/thumb/S104x104/?fname=https%3A%2F%2Ft1.daumcdn.net%2Fgift%2Fhome%2Ftheme%2F292020231106_MXMUB.png',
+  },
+];
 
 export const SelectTheme = () => {
   return (
-    <div>
-      <ThemeGrid />
+    <div className="theme-div">
+      {themes.map((theme) => (
+        <Link to="/theme/:themeKey">
+          <ThemeItem className="themeitem" title={theme.title} imgUrl={theme.imgUrl} />
+        </Link>
+      ))}
     </div>
   );
 };
